live-transmission/requirements.txt
future==1.0.0
iso8601==2.1.0
pyserial==3.5
PyYAML==6.0.1
# PyQt5
PyQt5
pyqtgraph

pandas
numpy
matplotlib

# scipy
scipy

# serial communication
pyserial==3.5

# Other libraries
scikit-learn
seaborn

